Move caas image repo changes to cmd not agent
As we need the details for the caas image repo from the host machine
we need to do it on the host machine not the agent. This is pure
insanity. Luckily, move it to here will at least prevent the double
switch once we're up and running.
